@@ -1,14 +1,16 @@
-import XCTest
+import Testing
+
 @testable import AdventOfCode
 
-final class Day1Tests: XCTestCase {
-    func testPart1Example() {
-        // Add your test cases here
-        XCTAssertEqual(Day1.solvePart1(), 0)
+@Suite("Day 1 Tests")
+struct Day1Tests {
+    @Test("Part 1 Example")
+    func part1Example() {
+        #expect(Day1.solvePart1() == 0)
     }
-    
-    func testPart2Example() {
-        // Add your test cases here
-        XCTAssertEqual(Day1.solvePart2(), 0)
+
+    @Test("Part 2 Example")
+    func part2Example() {
+        #expect(Day1.solvePart2() == 0)
     }
 }
